Fix installation routines.

+++ install_or_update.sh
@@ -31,7 +31,7 @@ self_update() {
         fi
 
         # we are on a user-defined branch
-        git fetch
+        git fetch origin
         if ! git rebase origin/master; then
             git rebase --abort
             exit_error "Could not rebase. Please manually run 'git rebase origin/master', resolve the conflicts and try again."
@@ -42,7 +42,7 @@ self_update() {
         fi
 
         # fresh install
-        git pull
+        git pull origin master
         git checkout -b "$default_user_branch"
         is_fresh_install="true"
 
@@ -98,7 +98,7 @@ new_backups="false"
 for i in "${!links[@]}"; do
     link="${links[$i]}"
     backup="${backups[$i]}"
-    if ! [[ -L $link ]]; then
+    if ! [[ -L $link ]] && [[ -f $link ]]; then
         cp "$link" "user-backup/$backup"
         new_backups="true"
     fi
@@ -114,46 +114,50 @@ for i in "${!links[@]}"; do
     target="${targets[$i]}"
     backup="${backups[$i]}"
     if ! [[ -L $link ]]; then
-        if [[ $link == ~/.bashrc ]]; then
-            read_reply "Note that I have backed up your '~/.bashrc' and will provide you with all the settings you would generally need." \
-                       "Nevertheless there may be some stuff you may want to keep. We will now search for this stuff to keep." \
-                       "Note that you can always review your original file content in '$(cd user-backup; pwd)/$backup' and add stuff to '$link' (or '$(pwd)/$target', resp.) later." \
-                       "I will now open your '~/.bashrc' in comparison with the original file of your distribution. The modified version will be on the right side." \
-                       "There you will search for the first (and optionally last) line numbers of the lines you want to keep." \
-                       "Press [Enter] to continue."
+        if [[ -f $link ]]; then
+            if [[ $link == ~/.bashrc ]]; then
+                read_reply "Note that I have backed up your '~/.bashrc' and will provide you with all the settings you would generally need." \
+                           "Nevertheless there may be some stuff you may want to keep. We will now search for this stuff to keep." \
+                           "Note that you can always review your original file content in '$(cd user-backup; pwd)/$backup' and add stuff to '$link' (or '$(pwd)/$target', resp.) later." \
+                           "I will now open your '~/.bashrc' in comparison with the original file of your distribution. The modified version will be on the right side." \
+                           "There you will search for the first (and optionally last) line numbers of the lines you want to keep." \
+                           "Press [Enter] to continue."
 
-            old_meld_conf=$(dconf read /org/gnome/meld/show-line-numbers)
-            dconf write /org/gnome/meld/show-line-numbers "true"
-            meld "/etc/skel/.bashrc" "$link"
-            dconf write /org/gnome/meld/show-line-numbers "$old_meld_conf"
+                old_meld_conf=$(dconf read /org/gnome/meld/show-line-numbers)
+                dconf write /org/gnome/meld/show-line-numbers "true"
+                meld "/etc/skel/.bashrc" "$link"
+                dconf write /org/gnome/meld/show-line-numbers "$old_meld_conf"
 
-            read_reply "Do you want to keep some lines now? [y|n]"
-            if [[ $REPLY =~ ^[Yy]$ ]]; then
-                read_reply "Enter first line to keep:"
-                start="$REPLY"
-                read_reply "Enter last line to keep (keep empty to select until the end of the file):"
-                end="$REPLY"
+                read_reply "Do you want to keep some lines now? [y|n]"
+                if [[ $REPLY =~ ^[Yy]$ ]]; then
+                    read_reply "Enter first line to keep:"
+                    start="$REPLY"
+                    read_reply "Enter last line to keep (keep empty to select until the end of the file):"
+                    end="$REPLY"
 
-                echo >> "$target"
-                if [[ ! -z "$end" ]]; then
-                    sed -n -e $start',$p' -e $end'q' "$link" >> "$target"
-                else
-                    sed -n $start',$p' "$link" >> "$target"
+                    echo >> "$target"
+                    if [[ ! -z "$end" ]]; then
+                        sed -n -e $start',$p' -e $end'q' "$link" >> "$target"
+                    else
+                        sed -n $start',$p' "$link" >> "$target"
+                    fi
                 fi
+            elif [[ $link == ~/.gitconfig ]]; then
+                cat "$link" >> "$target"
+            elif [[ $link == ~/.hstr_blacklist ]]; then
+                cp "$link" "$target"
+            elif [[ $link == ~/.hstr_favorites ]]; then
+                cp "$link" "$target"
+            elif [[ $link == ~/.vimrc ]]; then
+                cat "$link" >> "$target"
+            elif [[ $link == ~/.ssh/config ]]; then
+                cp "$link" "$target"
+            else
+                1>&2 echo "Implementation error: No installation routine for '$link'."
+                continue
             fi
-        elif [[ $link == ~/.gitconfig ]]; then
-            cat "$link" >> "$target"
-        elif [[ $link == ~/.hstr_blacklist ]]; then
-            cp "$link" "$target"
-        elif [[ $link == ~/.vimrc ]]; then
-            cat "$link" >> "$target"
-        elif [[ $link == ~/.ssh/config ]]; then
-            cp "$link" "$target"
-        else
-            1>&2 echo "Implementation error: No installation routine for '$link'."
-            continue
+            rm "$link"
         fi
-        rm "$link"
         ln -s "$(pwd)/$target" "$link"
         new_configs="true"
     fi
@@ -164,7 +168,13 @@ fi
 
 grep -q '^\. "\$HOME/\.biberconf/config/profile.sh"$' ~/.profile || echo -e '. "$HOME/.biberconf/config/profile.sh"\n' >> ~/.profile
 
+if ! [[ -d ~/.hstr_histories ]]; then
+    mkdir ~/.hstr_histories
+    if [[ -f ~/.bash_history ]]; then
+        cp ~/.bash_history ~/.hstr_histories/$(date -u +%Y-%m-%d_%H-%M-%S)_old-history
+    fi
+fi
+
 echo_success "Installation successfull." \
              "You should restart your terminal now or at least call 'source ~/.bashrc'." \
-             "You can always review and change the files in '$(pwd)' and commit them to your newly created branch 'custom' to keep track of your own changes." \
-             "If there are remote updates to biberconf, just call './install.sh' again to fetch the latest changes, install them and rebase your 'custom' branch."
+             "You can always review and change the files in '$(pwd)' and commit them to your newly created branch 'custom' to keep track of your own changes."
